+++ fs/sysfs.c
@@ -97,14 +97,14 @@ static int sysfs_readdir(void *fs_private, const char *path, vfs_dirent_t *entri
     }
 
     int count = subsystem_get_count();
-    int path_len = strlen(path);
+    size_t path_len = strlen(path);
 
     for (int i = 0; i < count && out < max; i++) {
         kernel_subsystem_t *s = subsystem_get_by_index(i);
         if (path_len == 0 || (strlen(s->name) > path_len && strncmp(s->name, path, path_len) == 0 && s->name[path_len] == '/')) {
             const char *sub_path = s->name + (path_len ? path_len + 1 : 0);
             char comp[64];
-            int j = 0;
+            size_t j = 0;
             while (sub_path[j] && sub_path[j] != '/' && j < 63) {
                 comp[j] = sub_path[j];
                 j++;
@@ -119,7 +119,7 @@ static int sysfs_readdir(void *fs_private, const char *path, vfs_dirent_t *entri
                 if (path_len == 0 || (strlen(ps->name) > path_len && strncmp(ps->name, path, path_len) == 0 && ps->name[path_len] == '/')) {
                     const char *p_sub_path = ps->name + (path_len ? path_len + 1 : 0);
                     char p_comp[64];
-                    int pj = 0;
+                    size_t pj = 0;
                     while (p_sub_path[pj] && p_sub_path[pj] != '/' && pj < 63) {
                         p_comp[pj] = p_sub_path[pj];
                         pj++;
@@ -178,7 +178,7 @@ static bool sysfs_exists(void *fs_private, const char *path) {
     }
 
     int count = subsystem_get_count();
-    int path_len = strlen(path);
+    size_t path_len = strlen(path);
     for (int i = 0; i < count; i++) {
         kernel_subsystem_t *s = subsystem_get_by_index(i);
         if (strlen(s->name) > path_len && strncmp(s->name, path, path_len) == 0 && s->name[path_len] == '/') return true;
